chore: rollback edit in dialog
